feature/create-tickets

+++ db/migrate/20241126154447_drop_events.rb
@@ -1,5 +1,5 @@
 class DropEvents < ActiveRecord::Migration[7.2]
   def change
-    drop_table :events
+    drop_table :events if table_exists?(:events)
   end
 end
